pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "seedream-mcp"
version = "1.0.0"
authors = [
    {name = "Seedream MCP Team", email = "tengmmvp@qq.com"}
]
description = "基于火山引擎 Seedream 4.0 API 的 MCP 工具，支持 AI 图像生成功能"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
keywords = ["mcp", "ai", "image-generation", "volcengine", "seedream", "stable-diffusion"]

dependencies = [
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.0.0",
    "volcengine-python-sdk[ark]>=1.0.0",
    "aiohttp>=3.8.0",
    "aiofiles>=23.0.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "Pillow>=10.0.0",
]

[tool.hatch.envs.dev]
dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "hatchling",
]

[project.urls]
Homepage = "https://github.com/caoergou/Seedream_MCP"
Documentation = "https://github.com/caoergou/Seedream_MCP#readme"
Repository = "https://github.com/caoergou/Seedream_MCP.git"
"Bug Tracker" = "https://github.com/caoergou/Seedream_MCP/issues"

[project.scripts]
seedream-mcp = "seedream_mcp.server:cli_main"

[tool.hatch.build.targets.sdist]
include = [
    "/seedream_mcp",
    "/pyproject.toml",
    "/README.md",
    "/LICENSE"
]

[tool.hatch.build.targets.wheel]
packages = ["seedream_mcp"]

[tool.hatch.envs.default]
dependencies = [
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.0.0",
    "volcengine-python-sdk[ark]>=1.0.0",
    "aiohttp>=3.8.0",
    "aiofiles>=23.0.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "Pillow>=10.0.0",
]

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"